Update NQCModels.jl
The the implementation, an example would be helpful of how to create a new type or abstract subtype. It's not entirely clear at the moment what is meant, especially, because the first sentence says not adding a subtype, but the last sentence says adding a subtype. I find it a bit confusing.