Add Subsystem and CompositeModel types
A CompositeModel combines multiple models assigned to determine parts
of the total system, e.g. different friction processes for different
parts of the system.

The Subsystem type is the building block of a CompositeModel, combining
an arbitrary NQCModels.Model with indices determining which particles
in the system it applies to.