Add Subsystems to apply models to selected atoms only (#39)
* Extend Harmonic for multiple particles and oscillators for tests

* Add Subsystem and CompositeModel types

A CompositeModel combines multiple models assigned to determine parts
of the total system, e.g. different friction processes for different
parts of the system.

The Subsystem type is the building block of a CompositeModel, combining
an arbitrary NQCModels.Model with indices determining which particles
in the system it applies to.

* Made ElectronicFrictionProviders a subtype of Model

This is so they are included as possible inputs for a Subsystem.

* Load CompositeModels

* CompositeModels are abstract Model sustypes

* Forgot to add dofs functions for CompositeModels

* Subsystem and CompositeModel have custom Base.show routines

* Subsystems should also accept colon indices

* Subsystem index logic now less type-constrained

* Diagnosing a StackOverflow with ASE models

* Possibly fixed type problems causing a stack overflow

Or not

Another try

Debug

Tried to break the stack overflow

* ElectronicFrictionProviders get dofs and ndofs too

* Attempt to fix friction matrix generation for subsystems

Need to call friction! and derivative! for subsystems and full R

I am very confused

* Made ASE interface aware of FixAtoms constraints

* Remove accidentally added ase PyCall extension with duplicate functionality

* Remove ASEconvert as no longer needed

* Ensure ase is available through CondaPkg

* Update docs

* Remove debug info

* Add unit test and fix small bug

* Naming inconsistency between dofs and ndofs

* CompositeModel derivative fixed.

* Version bump #39